+++ noodles-sam/src/header/parser.rs
@@ -27,7 +27,7 @@ pub enum ParseError {
     /// A read group ID is duplicated.
     DuplicateReadGroupId(Vec<u8>),
     /// A program ID is duplicated.
-    DuplicateProgramId(String),
+    DuplicateProgramId(Vec<u8>),
     /// A comment record is invalid.
     InvalidComment,
 }
@@ -60,7 +60,10 @@ impl fmt::Display for ParseError {
                 let id = str::from_utf8(buf).map_err(|_| fmt::Error)?;
                 write!(f, "duplicate read group ID: {id}")
             }
-            Self::DuplicateProgramId(id) => write!(f, "duplicate program ID: {id}"),
+            Self::DuplicateProgramId(buf) => {
+                let id = str::from_utf8(buf).map_err(|_| fmt::Error)?;
+                write!(f, "duplicate program ID: {id}")
+            }
             Self::InvalidComment => f.write_str("invalid comment record"),
         }
     }
@@ -350,7 +353,7 @@ mod tests {
 ";
         assert_eq!(
             parse(s),
-            Err(ParseError::DuplicateProgramId(String::from("pg0")))
+            Err(ParseError::DuplicateProgramId(Vec::from("pg0")))
         );
     }