+++ noodles-sam/CHANGELOG.md
@@ -49,8 +49,8 @@
   * sam/alignment/record_buf: Change quality scores to raw scores
     (`record_buf::QualityScores`).
 
-  * sam/header: Change `ReferenceSequences` and `ReadGroup` keys to byte
-    strings (`Vec<u8>`).
+  * sam/header: Change `ReferenceSequences`, `ReadGroup`, and `Program` keys to
+    byte strings (`Vec<u8>`).
 
   * sam/io/reader: Rename record to record buf and lazy record to record.